refactor/chore test: move setup fixtures to setup subfolder.
